Added todo
Former-commit-id: 34d79755f15cccc2a19ace029a86b071cd56b1a9 [formerly 4a7263d5213eaeacd7aa0fd788c71cc9969ff9c8] [formerly b164e9aa5282ce59820cd93d0509f638c86e24aa [formerly 3162c71c545fa342d2621bf7d49b48d2185678e1] [formerly d72b4231c29fba86840f990f1a08cd96271792a3] [formerly 8dce4bb6b9ca3f0312d5eb9154751f3de4048c24 [formerly 3852e16ccb7ec258803a0810fcca500864e233f9]] [formerly b164e9aa5282ce59820cd93d0509f638c86e24aa [formerly 3162c71c545fa342d2621bf7d49b48d2185678e1] [formerly d72b4231c29fba86840f990f1a08cd96271792a3] [formerly 8dce4bb6b9ca3f0312d5eb9154751f3de4048c24 [formerly 3852e16ccb7ec258803a0810fcca500864e233f9]] [formerly b164e9aa5282ce59820cd93d0509f638c86e24aa [formerly 3162c71c545fa342d2621bf7d49b48d2185678e1] [formerly d72b4231c29fba86840f990f1a08cd96271792a3] [formerly 8dce4bb6b9ca3f0312d5eb9154751f3de4048c24 [formerly 3852e16ccb7ec258803a0810fcca500864e233f9]] [formerly 2a4182f81d3e8cf10cc054a202a3369d83f16d3c [formerly e963e4e725c140db6a77261bddb85284b8266027]]]]]
Former-commit-id: e5af76df55f26df07247945ab6cb866ab92eb905
Former-commit-id: d84c010622a51a0503e347b8aeca88b6734c058a
Former-commit-id: 34c17ccc6cabf343e597cb407d163cbbe4f4048b
Former-commit-id: 62e487bc3b91a29848d0d28b5a60bd752c35aeeb
Former-commit-id: cfe2e357ee87029a27d5ae179514cf017c2094e3 [formerly 939af39b55e9cbde5f97ec08ec8116fa8feb75fd]
Former-commit-id: 4918e70f70a4fe0c2ac10817fb52e9ba24a6d0ba
Former-commit-id: 991c1597892ea9c2bb9fd26bdfe8c6c1d6bb4317

+++ Source/RawParserUWP/RawParserUWP/View/Pages/MainPage.xaml.cs
@@ -143,14 +143,14 @@ namespace RawParserUWP
                     //rawImage.CopyToBuffer(bitmap.PixelBuffer);
 
                     // test
-                    byte[] temp = currentRawImage.getImageAsByteArray();
+                    ushort[] temp = currentRawImage.getImageAsByteArray();
                     var bitmapContext = bitmap.GetBitmapContext();
 
                     for (int x = 0; x < currentRawImage.width; x++)
                     {
                         for (int y = 0; y < currentRawImage.height; y++)
                         {
-                            bitmap.SetPixel(x, y, temp[x + (y * currentRawImage.width)], temp[x + (y * currentRawImage.width)], temp[x + (y * currentRawImage.width)]);
+                            bitmap.SetPixel(x, y,(byte) temp[x + (y * currentRawImage.width)], (byte)temp[x + (y * currentRawImage.width)], (byte)temp[x + (y * currentRawImage.width)]);
                         }
                     }
                     bitmapContext.Dispose();
